+++ packages/react/runtime/src/lynx.ts
@@ -53,8 +53,13 @@ if (typeof __ALOG__ !== 'undefined' && __ALOG__) {
 
 if (__BACKGROUND__) {
   // Trick Preact and TypeScript to accept our custom document adapter.
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-  options.document = document as any;
+  options.document = document as unknown as Document;
+  if (lynx.queueMicrotask) {
+    options.requestAnimationFrame = callback => lynx.queueMicrotask(callback);
+  } else if (globalThis.Promise) {
+    const realResolvedPromise = globalThis.Promise.resolve();
+    options.requestAnimationFrame = callback => void realResolvedPromise.then(callback);
+  }
   setupBackgroundDocument();
   injectTt();
   addCtxNotFoundEventListener();