+++ packages/react/runtime/src/snapshot/spread.ts
@@ -46,11 +46,12 @@ function updateSpread(
   oldValue: Record<string, unknown> | undefined | null,
   elementIndex: number,
 ): void {
+  oldValue ??= {};
   let newValue: Record<string, unknown> = snapshot.__values![index] as Record<string, unknown>; // compiler guarantee this must be an object;
 
   const list = snapshot.parentNode;
   if (list?.__snapshot_def.isListHolder) {
-    const oldPlatformInfo = oldValue ? pick(oldValue, platformInfoAttributes) : undefined;
+    const oldPlatformInfo = pick(oldValue, platformInfoAttributes);
     const platformInfo = pick(newValue, platformInfoAttributes);
     if (!isDirectOrDeepEqual(oldPlatformInfo, platformInfo)) {
       if (__pendingListUpdates.values) {
@@ -75,8 +76,6 @@ function updateSpread(
       updateListItemPlatformInfo(fakeSnapshot, index, oldPlatformInfo, elementIndex);
     }
   }
-
-  oldValue ??= {};
 
   if (!snapshot.__elements) {
     return;