+++ packages/rspeedy/plugin-react-alias/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lynx-js/react-alias-rsbuild-plugin
 
+## 0.10.9
+
+### Patch Changes
+
+- Add alias for `@lynx-js/react/compat`. ([#1316](https://github.com/lynx-family/lynx-stack/pull/1316))
+
 ## 0.10.8
 
 ### Patch Changes