+++ packages/web-platform/web-core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @lynx-js/web-core
 
+## 0.15.3
+
+### Patch Changes
+
+- fix: improve compatibility with legacy template ([#1337](https://github.com/lynx-family/lynx-stack/pull/1337))
+
+  avoid "object Object" error for old version rspeedy outputs
+
+- Updated dependencies [[`0da5ef0`](https://github.com/lynx-family/lynx-stack/commit/0da5ef03e41f20e9f8019c6dc03cb4a38ab18854)]:
+  - @lynx-js/web-constants@0.15.3
+  - @lynx-js/web-mainthread-apis@0.15.3
+  - @lynx-js/web-worker-runtime@0.15.3
+  - @lynx-js/web-worker-rpc@0.15.3
+
 ## 0.15.2
 
 ### Patch Changes